10 min

+++ calculator.cpp
@@ -2,12 +2,12 @@
  using namespace std;
  
  int main(){
-    int X,Y;
+    long long X,Y;
     cin>>X>>Y;
  
-    int summation = X+Y;
-    int multiplication =X*Y;
-    int substraction= X-Y;
+    long long summation = X+Y;
+    long long multiplication = X*Y;
+    long long substraction= X-Y;
  
     cout<<X<<" + "<<Y<<" = "<<summation<<endl;
     cout<<X<<" * "<<Y<<" = "<<multiplication<<endl;
